Messages in prox are now saved in console
+Bug fixes

+++ KadTextChat/Config.cs
@@ -6,7 +6,7 @@ public class Config
     public int maxMessageLength { get; set; } = 80;
     public int maxWhisperLength { get; set; } = 20;
     public int maxYellLength { get; set; } = 40;
-    public float messageSpamCap { get; set; } = 4;
+    public int messageSpamCap { get; set; } = 4;
     public string speakSFX { get; set; } = "talking";
     public string speak0492SFX { get; set; } = "talkingzombie";
     public string speak049SFX { get; set; } = "talkingdoctor";
