+++ KadTextChat/TextToyInfoStore.cs
@@ -9,7 +9,6 @@ public sealed class TextToyInfoStore
     public AudioPlayer audioPlayer = null;
     public Speaker audioSpeaker = null;
     public TextComponent textComponent;
-    //Made as a float so it takes longer to decrement than increment
-    public float recentMessages = 0;
+    public int recentMessages = 0;
 }
